+++ packages/core/src/components/ozon-content/ozon-content-context.interface.ts
@@ -1,14 +1,18 @@
 import { EventEmitter } from "@stencil/core";
 import { OzonContentNodeState } from "./ozon-content-node-state.interface";
 
-import { OzonContentAnchorClickEvent, OzonContentMarkFunction } from "./ozon-content.interfaces";
+import {
+  OzonContentAnchorClickEvent,
+  OzonContentMarkFunction,
+  OzonContentUrlResolver,
+} from "./ozon-content.interfaces";
 
 export interface OzonContentContext {
   state: OzonContentNodeState;
-  addSpaceBeforeNode: boolean;
   inline: boolean;
   mark: OzonContentMarkFunction | undefined;
   setState(state: OzonContentNodeState): void;
   emitAnchorClick: EventEmitter<OzonContentAnchorClickEvent>["emit"];
   emitMarkItemHighlight(text: string, ref: HTMLElement): void;
+  urlResolver: OzonContentUrlResolver | undefined;
 }