+++ storybook/src/components/ozon-content/ozon-content.core-template.ts
@@ -8,12 +8,21 @@ export const coreOzonContent: ComponentImplementation<OzonContent> = {
   component: "ozonContent",
   implementation: "core",
   template: () =>
-    function ozonContentTemplate({ slotName, content, inline, mark, dsoAnchorClick, dsoOzonContentMarkItemHighlight }) {
+    function ozonContentTemplate({
+      slotName,
+      content,
+      inline,
+      mark,
+      urlResolver,
+      dsoAnchorClick,
+      dsoOzonContentMarkItemHighlight,
+    }) {
       return html`
         <dso-ozon-content
           slot=${ifDefined(slotName)}
           .content=${content}
           .mark=${ifDefined(mark)}
+          .urlResolver=${ifDefined(urlResolver)}
           ?inline=${inline}
           @dsoAnchorClick=${ifDefined(dsoAnchorClick)}
           @dsoOzonContentMarkItemHighlight=${ifDefined(dsoOzonContentMarkItemHighlight)}